Upgraded TemporalQuery to replace version 1.0 with 2.0
